predict_machine_maintenance/requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
imbalanced-learn==0.10.1
matplotlib==3.7.1
seaborn==0.12.2
joblib==1.3.2
requests==2.31.0
openpyxl==3.1.2

# Web framework
Django==4.2
djangorestframework==3.14.0
